+++ src/infrastructure/rendering.h
@@ -6,7 +6,7 @@
 typedef struct {
     int windowWidth;
     int windowHeight;
-    double aperture;
+    float aperture;
     gbVec3 position;
     gbVec3 view;
     gbVec3 up;
